requirements.txt
# PRA COREP Reporting Assistant - Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# RAG
sentence-transformers>=2.2.0
chromadb>=0.4.0
rank-bm25>=0.2.2
numpy>=1.24.0

# LLM
openai>=1.0.0

# App / UI
streamlit>=1.28.0
httpx>=0.25.0

# Utils
python-multipart>=0.0.6
